refactor(splunk): unexport ValidateCloudStack (#894)
ValidateCloudStack was exported temporarily during PR 1 (#55) of
the splunk HEC output rollout to let a black-box test exercise the
regex independently. The original godoc said exactly that:
"exposed now so PR 1 can test the regex independently". The
test-driven export was never a design decision.

Unexport it: rename to validateCloudStack, move the test into a
white-box file (splunk/config_internal_test.go) in package splunk
that calls the function directly. The internal caller in
parseSplunkCloudURL switches to the lowercase name.

This matches the loki precedent (validLabelName is a private
package var; no exported validation helper). Consumers reach the
same gate by calling Config.Validate (automatic on splunk.New);
the function never had an independent use case.

Refs #894
